Refactoring of the xpatch code into a separate file ##diff
